+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Version 1.8.1
+## Version 1.8.1 (11-03-2025)
 
 ### Added
 - Added `labels` to `stimplot` in `plotting`